Simplify build setup
